Changed the Human-Based-Compression directory and its children to follow proper Python style
